首页

+++ database/migrations/2019_06_03_174602_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->string('no')->unique()->comment('订单号');
             $table->unsignedBigInteger('user_id')->comment('用户ID');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->decimal('total_amount')->comment('金额');
+            $table->unsignedInteger('total_amount')->comment('金额');
             $table->dateTime('paid_at')->nullable()->comment('支付时间');
             $table->string('payment_method')->nullable()->comment('支付方式');
             $table->string('payment_no')->nullable()->comment('支付方订单号');
